+++ src/js/index.js
@@ -1,7 +1,7 @@
 'use strict'
 
 import Vue from 'vue';
-import App from './App.vue'
+import App from '../component/App.vue'
 
 new Vue({
   render: h => h(App),
