+++ webpack.config.js
@@ -6,7 +6,7 @@ const { VueLoaderPlugin } = require('vue-loader')
 
 module.exports = {
   mode: 'development',
-  entry: './src/index.js',
+  entry: './src/js/index.js',
   output: {
     path: `${outputPath}/js`,
     filename: 'bundle.js'
